+++ python_scripts/run_flash_demo_live.py
@@ -141,7 +141,7 @@ while True:
 
         check_faces.append(face)
         face, lmarks = face_processing.resize_face(face, bbx_)
-        facen = face_processing.get_normalized_face(face, lmarks.astype(np.int).reshape(1, 5, 2), face=True)
+        facen = face_processing.get_normalized_face(face, lmarks.astype(np.int32).reshape(1, 5, 2), face=True)
 
         cropped_aligned_faces.append(facen)
         cropped_aligned_faces.append(facen[:, ::-1, :])
